Change CustomSettingsPackage
