+++ Packages/UGF.CustomSettings/Runtime/CustomSettingsPackage.Editor.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: cfe91a27f0014317b3fe4dc6d36d1eba
+timeCreated: 1605034036
